Redirect PackageUtil consumers to langserver-core
Replace model-generator-commons' PackageUtil facade with direct
imports to langserver-core's PackageUtil across all consumer
modules. The facade is slimmed down to only resolveModulePackage(),
which remains to avoid a cyclic dependency with diagram-util.

- Slim model-generator-commons PackageUtil to a thin facade
- Update architecture-model-generator to import langserver-core
- Update flow-model-generator to import langserver-core
- Update service-model-generator to import langserver-core
- Adapt API calls to use langserver-core's method signatures
