+++ flow-model-generator/modules/flow-model-index-generator/src/main/java/io/ballerina/indexgenerator/SearchIndexGenerator.java
@@ -29,10 +29,10 @@ import io.ballerina.compiler.api.symbols.Qualifier;
 import io.ballerina.compiler.api.symbols.Symbol;
 import io.ballerina.compiler.api.symbols.TypeDefinitionSymbol;
 import io.ballerina.modelgenerator.commons.CommonUtils;
-import io.ballerina.modelgenerator.commons.PackageUtil;
 import io.ballerina.projects.Module;
 import io.ballerina.projects.ModuleDescriptor;
 import io.ballerina.projects.Package;
+import org.ballerinalang.langserver.common.utils.PackageUtil;
 
 import java.io.FileReader;
 import java.io.IOException;
@@ -67,6 +67,7 @@ public class SearchIndexGenerator {
     private static final String CONNECTOR_EXCLUDE_JSON = "connector_exclude.json";
 
     public static void main(String[] args) {
+        PackageUtil.initialize(false);
         SearchDatabaseManager.createDatabase();
 
         Gson gson = new Gson();
