+++ model-generator-commons/src/main/java/io/ballerina/modelgenerator/commons/AbstractLSTest.java
@@ -82,6 +82,7 @@ public abstract class AbstractLSTest {
         }
         log = LoggerFactory.getLogger(clazz());
         this.languageServer = new BallerinaLanguageServer();
+        this.languageServer.setPackageUtilOffline(true);
         TestUtil.LanguageServerBuilder builder = TestUtil.newLanguageServer().withLanguageServer(languageServer);
         this.serviceEndpoint = builder.build();
     }
